requirements.txt
# Core dependencies
torch>=1.10.0
torchvision>=0.11.0
ultralytics>=8.0.0  # For YOLOv8
numpy>=1.20.0
pyyaml>=6.0
matplotlib>=3.5.0

# Bayesian optimization
bayesian-optimization>=1.2.0

# Visualization and logging
tensorboard>=2.9.0

# ONNX Runtime for quantized model inference
onnx>=1.10.0
onnxruntime>=1.10.0
onnx-simplifier>=0.3.10

# Optional: GPU support for ONNX Runtime
# onnxruntime-gpu>=1.10.0

# Utilities
pillow>=9.0.0
tqdm>=4.62.0
pandas>=1.3.0
seaborn>=0.11.0
scikit-learn>=1.0.0

# Optional: For development
# pytest>=6.2.5
# black>=22.3.0
# isort>=5.10.0
# flake8>=4.0.0
